requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
filelock==3.12.3
frozenlist==1.4.0
fsspec==2023.9.0
huggingface-hub==0.17.1
idna==3.4
multidict==6.0.4
numpy==1.25.2
openai==0.28.0
packaging==23.1
PyPDF2==3.0.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
tiktoken==0.5.1
tokenizers==0.13.3
tqdm==4.66.1
transformers==4.33.1
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
trl @ git+https://github.com/huggingface/trl.git@7630f877f91c556d9e5a3baa4b6e2894d90ff84c